requirements.txt
Django>=1.9,<1.10
django-allauth==0.25.2
django-cors-headers==1.1.0
django-environ==0.4.0
django-grappelli==2.8.1
djangorestframework==3.3.3
djangorestframework-bulk==0.2.1
djangorestframework-jwt==1.8.0
django-rest-swagger==0.3.7
django-storages==1.4.1
django-extensions==1.6.7
django-phonenumber-field==1.1.0
django-rest-auth==0.8.1
django-timezone-field==1.3
Pillow==3.1.1
psycopg2==2.6.1
uWSGI==2.0.13.1
whitenoise==3.2
django-filter==0.15.0
python-dateutil==2.5.3
djangorestframework-filters==0.8.0
django-multiselectfield==0.1.4
tablib==0.11.2
eventlet==0.19.0
flower==0.9.1
pydot==1.1.0
